requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn
tqdm>=4.60.0
openai>=1.0.0
google-generativeai>=0.0.1
# deepseek>=0.0.1
requests